requirements.txt
fastapi
uvicorn
redis
pydantic
python-dotenv
PyJWT
aioredis
passlib
PyPDF2
Pillow
python-mimeparse
pikepdf
python-magic
opencv-python
paddlepaddle
paddleocr
pdf2image
reportlab
